Learning Project
This is a project where user can get  and as many learning resources as they want for reference purposes and also have access to them. it was implemented using vue.js. there are various functionalites including provide and inject, the use of slot, custom events, fragments, dynamic components, teleporting etc.